# .schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "runtime",
    "description": "The spiced runtime",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.2.0-unstable"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server. Configure with `--http`."
    }
  ],
  "paths": {
...
    "/v1/models": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "List Models",
        "description": "List all models, both machine learning and language models, available in the runtime.",
        "operationId": "get_models",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "description": "The format of the response (e.g., `json` or `csv`).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Format"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "If true, includes the status of each model in the response.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of models in JSON format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIModelResponse"
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "id": "gpt-4",
                      "object": "model",
                      "owned_by": "openai",
                      "datasets": null,
                      "status": "ready"
                    },
                    {
                      "id": "text-embedding-ada-002",
                      "object": "model",
                      "owned_by": "openai-internal",
                      "datasets": [
                        "text-dataset-1",
                        "text-dataset-2"
                      ],
                      "status": "ready"
                    }
                  ]
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                },
                "example": "\nid,object,owned_by,datasets,status\ngpt-4,model,openai,,ready\ntext-embedding-ada-002,model,openai-internal,\"text-dataset-1,text-dataset-2\",ready\n"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred while processing models",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "error": "App not initialized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentStatus": {
        "type": "string",
        "description": "Represents the status of a component (e.g. dataset, model, etc).",
        "enum": [
          "Initializing",
          "Ready",
          "Disabled",
          "Error",
          "Refreshing",
          "ShuttingDown"
        ]
      },
...
      "Format": {
        "type": "string",
        "enum": [
          "json",
          "csv"
        ]
      },
...
      "OpenAIModel": {
        "type": "object",
        "required": [
          "id",
          "object",
          "owned_by"
        ],
        "properties": {
          "datasets": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "The datasets associated with this model, if any"
          },
          "id": {
            "type": "string",
            "description": "The name of the model"
          },
          "object": {
            "type": "string",
            "description": "The type of the model (always `model`)"
          },
          "owned_by": {
            "type": "string",
            "description": "The source from which the model was loaded (e.g., `openai`, `spiceai`)"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentStatus",
                "description": "The status of the model (e.g., `ready`, `initializing`, `error`)"
              }
            ]
          }
        }
      },
      "OpenAIModelResponse": {
        "type": "object",
        "required": [
          "object",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIModel"
            }
          },
          "object": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}